implemented @ppkarwasz & @vy chamges regarding `FixedDateFormat.FixedFormat`
